move & clean functions from treeViewFunctions.js
